requirements.txt
bcrypt~=4.0.1
sqlalchemy~=2.0.40
werkzeug~=3.1.3
fastapi~=0.115.12
passlib~=1.7.4
PyJWT~=2.10.1
PyYAML~=6.0.2
pydantic~=2.11.4
uvicorn~=0.34.2
ngrok~=1.4.0
python-multipart~=0.0.20
pillow~=11.2.1